Moved Dao, Database, and TypeConverters files to database package
