+++ app/src/main/java/com/bignerdranch/android/androidreceiptapp/database/ReceiptDatabase.kt
@@ -1,8 +1,10 @@
-package com.bignerdranch.android.androidreceiptapp
+package com.bignerdranch.android.androidreceiptapp.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
+import com.bignerdranch.android.androidreceiptapp.Receipt
+import com.bignerdranch.android.androidreceiptapp.ReceiptEntry
 
 @Database(entities = [ Receipt::class, ReceiptEntry::class ], version=1, exportSchema = false)
 @TypeConverters(ReceiptTypeConverters::class)
